Extract AppToolbar and HeadSection components from App component
